+++ src/shared/hooks/mutate/index.ts
@@ -1,3 +1,4 @@
 export * from './product.mutate'
 export * from './product-detail.mutate'
 export * from './user.mutate'
+export * from './review.mutate'
