+++ src/shared/hooks/query/index.ts
@@ -2,3 +2,4 @@ export * from './product-detail.query'
 export * from './product.query'
 export * from './user.query'
 export * from './ranking.query'
+export * from './review.query'
