+++ src/shared/hooks/query/product-detail.query.ts
@@ -1,6 +1,6 @@
-import { useQuery, useSuspenseInfiniteQuery } from '@tanstack/react-query'
-import { fetchProductDetail, fetchProductsReviews } from '@shared/api'
-import { FetchProductsReviews, ProductDetailResponse } from '@shared/types'
+import { useQuery } from '@tanstack/react-query'
+import { fetchProductDetail } from '@shared/api'
+import { ProductDetailResponse } from '@shared/types'
 
 export function useFetchProductDetail(
   productId: number,
@@ -10,19 +10,5 @@ export function useFetchProductDetail(
     queryKey: ['product-detail', productId],
     queryFn: () => fetchProductDetail({ productId }),
     initialData,
-  })
-}
-
-export function useFetchProductReview({
-  productId,
-  order = 'recent',
-}: FetchProductsReviews) {
-  return useSuspenseInfiniteQuery({
-    queryKey: ['product-detail-review', productId, order],
-    queryFn: ({ pageParam }) =>
-      fetchProductsReviews({ productId, cursor: pageParam, order }),
-    initialPageParam: 0,
-    getNextPageParam: (lastPage) => lastPage.nextCursor || null,
-    select: (data) => (data.pages ?? []).flatMap((page) => page.list),
   })
 }
